Removed setAlignmentLimits from IAlignment
There is no reliable way to serialize the passed in IAlignmentLimit. Therefore there can't be any sort of "tool" that does this without the creator of multiblock controller being aware of what this tool will do. In that case, the programmer can just invent a private API.
It receive no usage in TecTech either way.