Drop private accessor in GuiScreenConfigureChannels
Inventory Bogo Sorter adds a mixin accessor to GuiScreen for buttonList. That worked fine on java 21, but resulted in an IllegalAccessException on java 8 in Inventory Bogo Sorter as the code tried to call this private method here in Structurelib rather than the mixin'd accessor.

This probably could also have been fixed in Inventory Bogo Sorter, however I also don't think a private accessor method makes a lot of sense.